requirements.txt
python-telegram-bot==21.6
anthropic==0.34.2
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.145.0
httpx==0.27.2
python-dotenv>=1.0.0
